tools/monGARS_deep_scan/requirements.txt
ruamel.yaml>=0.18
dockerfile-parse>=2.0
beautifulsoup4>=4.12
langdetect>=1.0.9
tqdm>=4.66
pytest>=7.4
